making some refactoring in the supervisor page and fix the image implementation

+++ back-end/src/main/java/com/example/backend/entities/IntershipOffre.java
@@ -30,7 +30,7 @@ public class IntershipOffre implements Serializable {
     private Boolean renumerete;
     private String company;
     private String supervisor;
-    private Long candidacy_number;
+    private int candidacy_number;
     private boolean saved;
     private Date creation_date;
     private String status;
